requirements.txt
pybit>=5.0.0
rich>=13.0.0
websocket-client>=1.6.0
pandas>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
